requirements.txt
accelerate==0.28.0
alabaster==0.7.13
annotated-types==0.6.0
antlr4-python3-runtime==4.8
anyio==4.2.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.3
bitarray==2.6.0
bleach==6.1.0
blessed==1.20.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.11.17
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloudpathlib==0.16.0
cmake==3.26.3
colorama==0.4.6
comm==0.2.1
confection==0.1.4
cycler==0.11.0
cymem==2.0.8
Cython==0.29.32
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.26.3
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.18.1
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.1
filelock==3.9.0
fonttools==4.34.4
fqdn==1.5.1
fsspec==2024.2.0
funcparserlib==1.0.0a0
gitdb==4.0.10
GitPython==3.1.31
gpustat==1.1.1
h11==0.14.0
huggingface-hub==0.21.4
hydra-core==1.0.7
idna==3.3
imagesize==1.4.1
importlib-metadata==6.6.0
importlib-resources==5.9.0
indic-nlp-library==0.91
ipykernel==6.29.0
ipython==8.12.3
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
kiwisolver==1.4.4
langcodes==3.4.0
language_data==1.2.0
lit==16.0.2
llvmlite==0.36.0
lxml==4.9.2
marisa-trie==1.1.1
MarkupSafe==2.1.2
matplotlib==3.5.2
matplotlib-inline==0.1.6
mistune==3.0.2
mock==5.0.2
Morfessor==2.0.6
mosestokenizer==1.2.1
mpmath==1.2.1
murmurhash==1.0.10
mypy-extensions==1.0.0
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
networkx==3.0rc1
notebook_shim==0.2.3
numpy==1.23.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.108
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
omegaconf==2.0.6
openfile==0.0.7
overrides==7.7.0
packaging==21.3
pandas==2.0.1
pandocfilters==1.5.1
parso==0.8.3
pathtools==0.1.2
pexpect==4.9.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
portalocker==2.7.0
preshed==3.0.9
prettytable==3.8.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==14.0.2
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.15.1
pyngrok==7.1.5
pyparsing==3.0.9
pyre-extensions==0.0.29
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-extension==0.2
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.2
referencing==0.33.0
regex==2023.5.5
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
sacrebleu==2.2.0
sacremoses==0.0.53
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.10.1
Send2Trash==1.8.2
sentencepiece==0.1.96
sentry-sdk==1.25.1
setproctitle==1.3.2
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==6.2.1
sphinx-argparse==0.4.0
sphinx-rtd-theme==1.2.0
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.36.3
subword-nmt==0.3.8
sympy==1.11.1
tabulate==0.9.0
tensorboardX==2.6
terminado==0.18.0
thinc==8.2.3
threadpoolctl==3.2.0
tinycss2==1.2.1
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolwrapper==2.1.0
torch==2.0.0
torchaudio==2.0.2
torchdata==0.6.0
torchtext==0.15.1
torchvision==0.15.2
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
transformers==4.38.2
triton==2.0.0
typer==0.9.4
types-python-dateutil==2.8.19.20240106
typing-inspect==0.8.0
typing_extensions==4.10.0
tzdata==2023.3
uctools==1.3.0
uri-template==1.3.0
urllib3==1.26.16
uvicorn==0.28.0
wandb==0.16.2
wasabi==1.1.2
wcwidth==0.2.6
weasel==0.3.4
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
zipp==3.17.0
